Create classes dir & education.ts data.
