+++ src/classes/Education.ts
