+++ src/components/routes.tsx
@@ -5,7 +5,7 @@ import SkillsContainer from "./skills";
 import Bio from "./bio";
 import Education from "./education";
 import Projects from "./projects";
-import type {Resume} from "../data/Resume";
+import type {Resume} from "../classes/Resume";
 
 export default function Router({resume}: {resume: Resume}) {
   return <Routes>
